Update PlaneTracker.cs
Trying to put bounds on the tracking function.  This version finds the closest neighbor polygon in frame ti and checks if its centroid is within the plane under consideration's polygon.  If it is, we consider it a match.  If not, we hold onto the plane at its current position and start a counter to kill the plane.  Slightly more robust than a global cut-off for distance, but this still might not be the best way.